pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "image-agent"
version = "0.1.0"
description = "AI Image Agent — generate, edit, upscale, remove bg, describe. Routes to any model."
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "pydantic>=2.7.0",
    "click>=8.0",
    "Pillow>=10.0.0",
    "python-multipart>=0.0.9",
    "prometheus-client>=0.20.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-anthropic>=0.3.0",
]

[project.optional-dependencies]
local = ["diffusers>=0.30.0", "torch>=2.0.0", "transformers>=4.40.0", "rembg>=2.0.0"]
dev = ["pytest", "pytest-asyncio", "respx", "httpx", "Pillow"]

[project.scripts]
image-agent = "image_agent.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/image_agent"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
